Feat/stock dashboard

+++ src/main/java/com/stockmanager/stockmanager_be/service/impl/CategoryServiceImpl.java
@@ -1,13 +1,17 @@
 package com.stockmanager.stockmanager_be.service.impl;
 
-import com.stockmanager.stockmanager_be.dto.CategoryRequestDto;
-import com.stockmanager.stockmanager_be.dto.CategoryResponseDto;
+import com.stockmanager.stockmanager_be.constant.CommonMessageConstant;
+import com.stockmanager.stockmanager_be.dto.request.CategoryRequestDto;
+import com.stockmanager.stockmanager_be.dto.response.CategoryResponseDto;
+import com.stockmanager.stockmanager_be.dto.response.ResponseEntityDto;
 import com.stockmanager.stockmanager_be.entity.Category;
 import com.stockmanager.stockmanager_be.exception.CategoryNotFoundException;
 import com.stockmanager.stockmanager_be.exception.DuplicateCategoryException;
 import com.stockmanager.stockmanager_be.mapper.CategoryMapper;
 import com.stockmanager.stockmanager_be.repo.CategoryRepo;
 import com.stockmanager.stockmanager_be.service.CategoryService;
+import com.stockmanager.stockmanager_be.type.ResponseStatus;
+import com.stockmanager.stockmanager_be.util.MessageUtil;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
@@ -20,24 +24,22 @@ public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepo categoryRepo;
     private final CategoryMapper categoryMapper;
+    private final MessageUtil messageUtil;
 
-    public CategoryServiceImpl(CategoryRepo categoryRepo, CategoryMapper categoryMapper) {
+
+    public CategoryServiceImpl(CategoryRepo categoryRepo, CategoryMapper categoryMapper, MessageUtil messageUtil) {
         this.categoryRepo = categoryRepo;
         this.categoryMapper = categoryMapper;
+        this.messageUtil = messageUtil;
     }
 
     @Override
-    public CategoryResponseDto createCategory(CategoryRequestDto categoryRequestDto) {
-        try {
-            Category category = categoryMapper.toCategory(categoryRequestDto);
+    public ResponseEntityDto createCategory(CategoryRequestDto categoryRequestDto) {
+        String message = messageUtil.getMessage(CommonMessageConstant.COMMON_SUCCESS_CATEGORY_CREATED);
 
-            Category newCategory = categoryRepo.save(category);
-            return categoryMapper.toCategoryResponseDto(newCategory);
-        } catch (DataIntegrityViolationException e){
-            throw new DuplicateCategoryException("Category with name '" + categoryRequestDto.getName() + "' already exists");
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to create category " + e.getMessage() );
-        }
+        Category category = categoryMapper.toCategory(categoryRequestDto);
+        Category savedCategory = categoryRepo.save(category);
+        return new ResponseEntityDto(ResponseStatus.SUCCESSFUL, message, categoryMapper.toCategoryResponseDto(savedCategory));
     }
 
     @Override
@@ -46,62 +48,53 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(Integer categoryId) {
-
-        try {
-            categoryRepo.deleteById(categoryId);
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to delete category " + e.getMessage() );
+    public ResponseEntityDto deleteCategory(Integer categoryId) {
+        String message = messageUtil.getMessage(CommonMessageConstant.COMMON_SUCCESS_CATEGORY_DELETED);
+        if (!categoryRepo.existsById(categoryId)) {
+            throw new CategoryNotFoundException(CommonMessageConstant.COMMON_ERROR_CATEGORY_NOT_FOUND);
         }
+        categoryRepo.deleteById(categoryId);
+        return new ResponseEntityDto(ResponseStatus.SUCCESSFUL,message);
     }
 
     @Override
-    public CategoryResponseDto getCategoryById(Integer categoryId) {
-        try {
-            Optional<Category> category = categoryRepo.findById(categoryId);
-            if (category.isPresent()) {
-                return categoryMapper.toCategoryResponseDto(category.get());
-            } else {
-                throw new CategoryNotFoundException("Category not found with id: " + categoryId);
-            }
-        } catch (CategoryNotFoundException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to get Category " + e.getMessage());
+    public ResponseEntityDto getCategoryById(Integer categoryId) {
+
+        String message = messageUtil.getMessage(CommonMessageConstant.COMMON_SUCCESS_DATA_RETRIEVED);
+        Optional<Category> category = categoryRepo.findById(categoryId);
+        if (category.isEmpty()){
+            throw new CategoryNotFoundException(CommonMessageConstant.COMMON_ERROR_CATEGORY_NOT_FOUND);
         }
+        CategoryResponseDto categoryResponseDto = categoryMapper.toCategoryResponseDto(category.get());
+        return new ResponseEntityDto(ResponseStatus.SUCCESSFUL, message, categoryResponseDto);
     }
 
     @Override
-    public List<CategoryResponseDto> getAllCategories() {
-        List<CategoryResponseDto> categoryList = new ArrayList<>();
-        try {
-            List<Category> categories = categoryRepo.findAll();
+    public ResponseEntityDto getAllCategories() {
+        String message = messageUtil.getMessage(CommonMessageConstant.COMMON_SUCCESS_LIST_RETRIEVED);
+        List<CategoryResponseDto> categoryList = categoryRepo.findAll()
+                .stream()
+                .map(categoryMapper::toCategoryResponseDto)
+                .toList();
 
-            for (Category category : categories) {
-                CategoryResponseDto categoryResponseDto = categoryMapper.toCategoryResponseDto(category);
-                categoryList.add(categoryResponseDto);
-            }
+        return new ResponseEntityDto(ResponseStatus.SUCCESSFUL, message, categoryList);
 
-            return categoryList;
-
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to get all categories " + e.getMessage() );
-        }
     }
 
     @Override
     public CategoryResponseDto findCategoryByName(String categoryName) {
-        try {
-            Optional<Category> category = categoryRepo.findByName(categoryName);
-            if (category.isPresent()) {
-                return categoryMapper.toCategoryResponseDto(category.get());
-            } else {
-                throw new CategoryNotFoundException("Category not found with name: " + categoryName);
-            }
-        } catch (CategoryNotFoundException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new RuntimeException("Failed to get Category " + e.getMessage());
-        }
+        return null;
+//        try {
+//            Optional<Category> category = categoryRepo.findByName(categoryName);
+//            if (category.isPresent()) {
+//                return categoryMapper.toCategoryResponseDto(category.get());
+//            } else {
+//                throw new CategoryNotFoundException("Category not found with name: " + categoryName);
+//            }
+//        } catch (CategoryNotFoundException e) {
+//            throw e;
+//        } catch (Exception e) {
+//            throw new RuntimeException("Failed to get Category " + e.getMessage());
+//        }
     }
 }
